+++ packages/core-browser/src/common/common.module.ts
@@ -6,19 +6,16 @@ import { DefaultOpenerContribution, OpenerContributionClient } from '../opener/o
 import { CommonServerPath, CryptrServicePath, KeytarServicePath } from '@opensumi/ide-core-common';
 import { AuthenticationContribution } from '../authentication/authentication.contribution';
 import { HashCalculateContribution } from '../hash-calculate/hash-calculate.contribution';
-import { RemoteOpenerConverterContributionClient } from '../remote-opener/converter.contribution';
-import { RemoteOpenerConverterContribution } from '../remote-opener';
 
 @Injectable()
 export class ClientCommonModule extends BrowserModule {
-  contributionProvider = [OpenerContribution, RemoteOpenerConverterContribution];
+  contributionProvider = [OpenerContribution];
   providers = [
     ClientCommonContribution,
     DefaultOpenerContribution,
     OpenerContributionClient,
     AuthenticationContribution,
     HashCalculateContribution,
-    RemoteOpenerConverterContributionClient,
   ];
   backServices = [
     {
