+++ packages/remote-opener/src/browser/index.ts
@@ -1,19 +1,20 @@
 import { Injectable } from '@opensumi/di';
 import { BrowserModule } from '@opensumi/ide-core-browser';
-import { RemoteOpenerBrowserServiceToken } from '@opensumi/ide-core-browser/lib/remote-opener';
-
-import { RemoteOpenerServicePath } from '../common';
+import { RemoteOpenerConverterContribution, RemoteOpenerServicePath, RemoteOpenerBrowserServiceToken } from '../common';
+import { RemoteOpenerContributionClient } from './remote.opener.contribution';
 import { RemoteOpenerBrowserServiceImpl } from './remote.opener.service';
 
 export * from './remote.opener.service';
 
 @Injectable()
 export class RemoteOpenerModule extends BrowserModule {
+  contributionProvider = [RemoteOpenerConverterContribution];
   providers = [
     {
       token: RemoteOpenerBrowserServiceToken,
       useClass: RemoteOpenerBrowserServiceImpl,
     },
+    RemoteOpenerContributionClient,
   ];
   backServices = [
     {
